Move track alignment functions to new source file.
